Fix task.

+++ lib/tasks/application.rake
@@ -49,12 +49,10 @@ namespace :application do
   desc "Cleanup expired proxy users"
   task :cleanup_proxy_users => :environment do
     ProxyUser.includes(:user).where("expired_at < ?", Time.zone.today).each do |proxy_user|
-      ProxyUser.transaction do
-        ProxyUserService.delete_proxy_user_in_alma(
-          proxy_user: proxy_user,
-          current_user: proxy_user.user
-        )
-
+      if ProxyUserService.delete_proxy_user_in_alma(
+        proxy_user: proxy_user,
+        current_user: proxy_user.user
+      )
         proxy_user.destroy!
       end
     end
